apps/worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
apscheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
prisma==0.13.0
boto3==1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
ruff==0.1.7
